Ignore rowid in SQLDatabaseController.merge_databases_new
`rowid` is built-in to sqlite tables and we use it extensively
when querying throughout the codebase.  It has been added to the
postgres tables as well but that causes code to fail because the code
doesn't expect `rowid` to be returned.  So we need to remove the `rowid`
column and adjust indices.