Define exception for an out-of-sync table
When the table defined in code differs from that in the database, this
will raise an error instead of deleting the table.

Seems like a not so good idea for a server application to delete it's
own data. I can understand desire from the GUI app perspective, but it
doesn't make sense for an unsupervised process.